move to amrexastro_cxx_network from base_cxx_network
